+++ src/components/form/ChangePwdNotLoggedInForm.js
@@ -32,7 +32,7 @@ function ChangePwdNotLoggedInForm() {
             if (memberId !== null) {
                 const response = await changePwdNotLoggedInAPI(memberId, form.newPassword, form.confirmNewPassword);
                 console.log("response: ", response);
-                if (response?.success) {
+                if (response.results?.success) {
                     localStorage.removeItem("memberId");
                     setWarningMessage('');
 
@@ -40,8 +40,8 @@ function ChangePwdNotLoggedInForm() {
                     setPopupOverlay(true);
                     setPopupMsg('비밀번호 변경 성공!');
                 } else {
-                    if (response.data && response.data.error) {
-                        setWarningMessage(response.data.error);
+                    if (response.data.results && response.data.results.error) {
+                        setWarningMessage(response.data.results.error);
                     } else {
                         setWarningMessage('비밀번호를 다시 확인해주세요.');
                     }
